requirements.txt
google-api-python-client==2.160.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
cloudinary==1.44.1
beautifulsoup4==4.12.3

